4_training_azure_ml/requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
sentencepiece>=0.1.99
accelerate>=0.20.0
azureml-core>=1.50.0
azureml-dataset-runtime>=1.50.0
azureml-defaults>=1.50.0
tensorboard>=2.13.0 